LaTeX space retention in glosses
Implements code for maintaining spaces in English glosses for use in LaTeX. This isn't the Leipzig glossing way, but the way we have the lexicon set up in the standard FLEx instance, none of the definitions are Leipzig-compatible anyway, so this is a good solution for that.

Also comments out isSuffix implementation as it doesn't work well for Tunica at the moment.